requirements.txt
apache-beam==2.61.0
apache-flink==2.2.0
apache-flink-libraries==2.2.0
attrs==26.1.0
avro==1.12.1
certifi==2026.4.22
charset-normalizer==3.4.7
cloudpickle==2.2.1
confluent-kafka==2.14.0
crcmod==1.7
dill==0.3.1.1
dnspython==2.8.0
docopt==0.6.2
fastavro==1.12.2
fasteners==0.20
find_libpython==0.5.1
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.13
jsonpickle==3.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.1.3
objsize==0.7.1
orjson==3.11.9
packaging==26.2
pandas==2.2.3
pemja==0.5.5
proto-plus==1.28.0
protobuf==5.29.6
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pydot==1.4.2
PyJWT==2.12.1
pymongo==4.17.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.2
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
regex==2026.5.9
requests==2.33.1
rpds-py==0.30.0
ruamel.yaml==0.19.1
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
zstandard==0.25.0
